+++ packages/terra-action-header/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Changed
 * Inline SVG aligned for IE and wdio screenshots Updated.
 
+### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
+
+2.20.0 - (July 16, 2019)
+------------------
 ### Added
 * Added documentation on required peerDependencies