+++ packages/terra-dialog/CHANGELOG.md
@@ -3,12 +3,16 @@ ChangeLog
 
 Unreleased
 ----------
+
+2.17.0 - (July 16, 2019)
+------------------
 ### Added
 * Added documentation on required peerDependencies
 
 2.16.0  - (July 11, 2019)
 ------------------
 ### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports
 * Updated Docs around the requirement of terra-base and aggregate-translations.